examples/requirements.txt
numpy
Pillow